feat: add UDM system, cost calculation and base-unit stock management
- Add UnitOfMeasure enum (G, KG, ML, L, UN, DUZIA) with conversion factors
- Add ProductType enum (INSUMO, PRODUTO_FINAL, PRODUTO_INTERMEDIARIO)
- Product: add tipo, purchaseUnit, packageQuantity fields; storage/minStorage -> BigDecimal
- TechnicalSheet: add rendimento, labourCost%, variableExpenses%, desiredMargin% fields
- TechnicalSheetItem: add unit (UnitOfMeasure) per item for recipe-level UDM
- StockMovement: quantity -> BigDecimal (base unit storage throughout)
- TechnicalSheetService: new calculateCost() with markup-divisor pricing algorithm
- PurchaseService: toBaseUnits() converts purchased packages to base units on stock entry
- ProductionOrderService: deducts ingredients in base units via item UDM conversion
- SaleService: stock validation and deduction now uses BigDecimal comparison
- New endpoint GET /technical-sheet/{id}/calculate-cost
- Frontend: ProductForm with tipo/UDM/packageQty fields and dynamic stock unit suffix
- Frontend: TechnicalSheetForm with per-item UDM select, pricing section, cost preview panel
- Frontend: ProductList badge shows stock with correct base unit (g/ml/un)

+++ SigRestBack/src/main/java/br/com/sigrest/api/dto/StockMovementRequestDTO.java
@@ -3,13 +3,14 @@ package br.com.sigrest.api.dto;
 import br.com.sigrest.api.entity.StockMovement;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class StockMovementRequestDTO {
     private Long productId;
     private StockMovement.MovementType type;
-    private Integer quantity;
+    private BigDecimal quantity;
     private LocalDateTime date;
     private String description;
 }
